requirements.txt
requests
python-dotenv
pyyaml
pytest
